+++ CHANGELOG.md
@@ -4,4 +4,5 @@
 * **New Resource:** `identitycloud_custom_domains` ([#8](https://github.com/pingidentity/terraform-provider-pingfederate/pull/8))
 * **New Resource:** `identitycloud_certificate` ([#9](https://github.com/pingidentity/terraform-provider-pingfederate/pull/9))
 * **New Resource:** `identitycloud_cookie_domains` ([#10](https://github.com/pingidentity/terraform-provider-pingfederate/pull/10))
+* **New Resource:** `identitycloud_certificate_signing_request_response` ([#19](https://github.com/pingidentity/terraform-provider-pingfederate/pull/19))
 * **New Resource:** `identitycloud_sso_cookie` ([#20](https://github.com/pingidentity/terraform-provider-pingfederate/pull/20))